Refactored classes into their own files
